more bug fixes to handle vote changes
